feat(pairing): latest tls lib
