+++ app/scripts/lib/constants.js
@@ -133,7 +133,6 @@ module.exports = {
 
   DEVICE_PAIRING_AUTHORITY_CONTEXT: 'device_pairing_authority',
   DEVICE_PAIRING_AUTHORITY_REDIRECT_URI: 'urn:ietf:wg:oauth:2.0:oob:pair-auth-webchannel',
-  DEVICE_PAIRING_CHANNEL_KEY_BYTES: 32,
   DEVICE_PAIRING_SCOPES: ['profile', 'https://identity.mozilla.com/apps/oldsync'],
   DEVICE_PAIRING_SUPPLICANT_CONTEXT: 'device_pairing_supplicant',
